Refactor: Change primary key types to Long and rename CommunicationLog
This commit updates the primary key types for several entities (ContactTierEntity, TagEntity, CommunicationLog, ContactEntity, ContactTagCrossRef) from Int to Long. This change provides a larger range for primary keys, preventing potential overflow issues as the database grows.

Additionally, the CommunicationLog entity has been renamed to CommunicationLogEntity to maintain consistency with other entity naming conventions. Corresponding updates have been made in tests and utility functions to reflect this renaming. The database version has been incremented to 4 to accommodate these schema changes.

+++ app/src/main/java/com/hollowvyn/kneatr/data/local/entity/ContactEntity.kt
@@ -16,10 +16,10 @@ import androidx.room.PrimaryKey
     ],
 )
 data class ContactEntity(
-    @PrimaryKey(autoGenerate = true) val contactId: Int,
+    @PrimaryKey(autoGenerate = true) val contactId: Long,
     val name: String,
     val phoneNumber: String,
     val email: String? = null,
-    val tierId: Int? = null,
+    val tierId: Long? = null,
     val customFrequencyDays: Int? = null,
 )
